+++ src/widgets/comments/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./comments"
